+++ dto/backchannel_authentication_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019 Authlete, Inc.
+// Copyright (C) 2019-2020 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -77,6 +77,9 @@ type BackchannelAuthenticationResponse struct {
 
 	//
 	RequestContext string `json:"requestContext,omitempty"`
+
+	//
+	Resources []string `json:"resources,omitempty"`
 
 	//
 	Warnings []string `json:"warnings,omitempty"`