+++ CHANGES.md
@@ -20,6 +20,9 @@ CHANGES
 - `BackchannelAuthenticationFailReason` type
     * Added `BackchannelAuthenticationFailReason_INVALID_TARGET`.
 
+- `BackchannelAuthenticationResponse` struct
+    * Added `Resources` field.
+
 v1.0.5 (2019-09-25)
 -------------------